pyproject.toml
[project]
name = "solo-mcp"
version = "0.1.0"
description = "Python Function Vector Store MCP Server (Solo Edition)"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.1.0",
    "numpy",
    "duckdb",
    "google-genai",     # Lightweight Embeddings
    "psutil",           # For process management
    "tqdm>=4.66.0",     # Fix for TMonitor import error
    "pyinstaller>=6.3.0",
    "docker>=7.0.0",
    "flet>=0.21.0",
    "keyring",          # For secure storage of API keys
    "ruff>=0.1.0",      # Quality Gate: Linter
    "mypy>=1.6.0",      # Quality Gate: Type Checker
    "pytest>=8.0.0",    # Testing
    "torch>=2.0.0",     # Translation & Quality Gate
    "transformers>=4.38.0",
    "bitsandbytes>=0.42.0", # 4-bit quantization
    "huggingface_hub>=0.20.0",
]

[project.scripts]
solo-mcp = "solo_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["solo_mcp"]
